requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
openai>=1.102.0
ddgs>=8.1.0
python-dotenv>=1.0.0
pytest>=7.0.0
